Revert "Friends"

+++ routes/controller/meal.js
@@ -17,10 +17,6 @@ const {
 
 async function getMeal(req, res, next) {
   try {
-    if (req.accessLevel !== ACCESS_LEVELS.CREATOR) {
-      throw createError(UNAUTHORIZED);
-    }
-
     const pagenateOptions = { ...defaultOption };
 
     const { userId } = req;
@@ -186,12 +182,14 @@ async function deleteMealDetail(req, res, next) {
       throw createError(NOT_FOUND);
     }
 
-    const imageKey = meal.url.split("/album1/").pop();
+    if (meal.url) {
+      const imageKey = meal.url.split("/album1/").pop();
 
-    await s3.deleteObject({
-      Bucket: process.env.BUCKET_NAME,
-      Key: `album1/${imageKey}`,
-    }).promise();
+      await s3.deleteObject({
+        Bucket: process.env.BUCKET_NAME,
+        Key: `album1/${imageKey}`,
+      }).promise();
+    }
 
     await meal.deleteOne();
     await Comment.deleteMany({ ratingId: meal._id });
